fix: incompatible data type in comments table

+++ database/migrations/2023_03_22_061824_create_comments_table.php
@@ -15,7 +15,7 @@ class CreateCommentsTable extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('post_id');
+            $table->unsignedBigInteger('post_id');
             $table->text('comment');
             $table->foreign('post_id')
                 ->references('id')
